Require Errors to be Debug + Display
One goal of this library is to provide a clean interface for other libraries. Errors need to be visible to be useful, it's common that they are required to be Debug + Display. Without these requirements here, they must be added to downstream libraries APIs which is gnarly. For example:

```rust
pub(crate) fn cached_layer_write_metadata<M, B>(
    layer_name: LayerName,
    context: &BuildContext<B>,
    metadata: &'_ M,
) -> libcnb::Result<LayerRef<B, Meta<M>, Meta<M>>, B::Error>
where
    B: libcnb::Buildpack,
    M: CacheDiff + TryMigrate + Serialize
    <M as TryMigrate>::Error: std::fmt::Display, // <=== HERE
{
    // ...
}
```

With this change applied it should allow applications to remove that extra qualification:

```rust
pub(crate) fn cached_layer_write_metadata<M, B>(
    layer_name: LayerName,
    context: &BuildContext<B>,
    metadata: &'_ M,
) -> libcnb::Result<LayerRef<B, Meta<M>, Meta<M>>, B::Error>
where
    B: libcnb::Buildpack,
    M: CacheDiff + TryMigrate + Serialize
    // Much cleaner now!
{
    // ...
}
```

This is an API change, so it's a major version rev. Though most (if not all) consumers would be using this with errors that are already Debug + Display.